Added user authentication and lots of bug fixes

+++ easy_barn/lib/barns_list_page.dart
@@ -3,8 +3,9 @@ import 'package:easy_barn/animal_class.dart';
 import 'package:easy_barn/barn_class.dart';
 import 'package:easy_barn/create_animal_form.dart';
 import 'package:easy_barn/create_barn_form.dart';
-import 'package:easy_barn/create_person_form.dart';
+import 'package:easy_barn/login/log_in_page.dart';
 import 'package:easy_barn/person_class.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import './animal_list_page.dart' as animalList;
@@ -36,31 +37,43 @@ class _BarnList extends State<BarnList> {
                   builder: (ctx) => animalList.AnimalList(name: barn.name)));
             },
             onLongPress: () async {
-              showDialog(
-                  context: context,
-                  builder: (BuildContext context) {
-                    return AlertDialog(
-                      title: const Text('Delete Barn'),
-                      content: const Text(
-                          'You are about to delete the selected Barn which will delete all animals'
-                          'within it and remove the barn from any boarders\' accounts.\n'
-                          'Is this really what you want?'),
-                      actions: [
-                        TextButton(
-                            onPressed: () async {
-                              await deleteBarn(barn);
-                              Navigator.pop(context);
-                              setState(() {});
-                            },
-                            child: const Text('Delete')),
-                        TextButton(
-                            onPressed: () {
-                              Navigator.pop(context);
-                            },
-                            child: const Text('Cancel'))
-                      ],
-                    );
-                  });
+              if (main.MyApp.currentUser.id == barn.ownerid) {
+                showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return AlertDialog(
+                        title: const Text('Delete Barn'),
+                        content: const Text(
+                            'You are about to delete the selected Barn which will delete all animals'
+                            'within it and remove the barn from any boarders\' accounts.\n'
+                            'Is this really what you want?'),
+                        actions: [
+                          TextButton(
+                              onPressed: () async {
+                                await deleteBarn(barn);
+                                Navigator.pop(context);
+                                setState(() {});
+                              },
+                              child: const Text('Delete')),
+                          TextButton(
+                              onPressed: () {
+                                Navigator.pop(context);
+                              },
+                              child: const Text('Cancel'))
+                        ],
+                      );
+                    });
+              } else {
+                showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return const AlertDialog(
+                        title: Text('Delete Barn'),
+                        content: Text(
+                            "If you wish to be removed from the barn, please contact the barn owner"),
+                      );
+                    });
+              }
             },
           ),
         );
@@ -70,14 +83,8 @@ class _BarnList extends State<BarnList> {
   Widget build(BuildContext ctx) {
     return Scaffold(
         appBar: AppBar(
-          backgroundColor: Color.fromARGB(255, 51, 91, 122),
-          title: Text(
+          title: const Text(
             "Easy Barn",
-            style: GoogleFonts.bitter(
-                textStyle: const TextStyle(
-                    color: Color.fromARGB(255, 244, 221, 177),
-                    fontSize: 28,
-                    fontWeight: FontWeight.w600)),
           ),
         ),
         endDrawer: NavigationDrawer(children: [
@@ -96,7 +103,7 @@ class _BarnList extends State<BarnList> {
                             fontWeight: FontWeight.w600)),
                   )))),
           ListTile(
-            title: const Text("Add New"),
+            title: const Text("Add New Barn or Animal"),
             onTap: () {
               showDialog(
                   context: context,
@@ -104,16 +111,9 @@ class _BarnList extends State<BarnList> {
                     return AlertDialog(
                         title: const Text('Add New'),
                         content: SizedBox(
-                            height: 200,
-                            child: Column(children: [
-                              TextButton(
-                                  onPressed: () async {
-                                    Navigator.of(ctx).push(MaterialPageRoute(
-                                        builder: (ctx) =>
-                                            const CreatePersonForm()));
-                                    setState(() {});
-                                  },
-                                  child: const Text('Add Person')),
+                            child: Column(
+                                mainAxisSize: MainAxisSize.min,
+                                children: [
                               TextButton(
                                   onPressed: () async {
                                     await Navigator.of(ctx).push(
@@ -121,6 +121,7 @@ class _BarnList extends State<BarnList> {
                                             builder: (ctx) =>
                                                 const CreateBarnForm()));
                                     setState(() {});
+                                    Navigator.of(ctx).pop();
                                   },
                                   child: const Text('Add Barn')),
                               TextButton(
@@ -130,10 +131,110 @@ class _BarnList extends State<BarnList> {
                                             builder: (ctx) =>
                                                 const CreateAnimalForm()));
                                     setState(() {});
+                                    Navigator.of(ctx).pop();
                                   },
                                   child: const Text('Add Animal'))
                             ])));
                   });
+            },
+          ),
+          Divider(
+            height: 1,
+            color: Colors.blueGrey.shade800,
+            thickness: 1,
+          ),
+          ListTile(
+              title: const Text("Join Existing Barn"),
+              onTap: () {
+                String barnToJoin = "";
+
+                showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return AlertDialog(
+                        title: const Text('Join Barn'),
+                        content: Column(
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            const Text(
+                                "Enter the invite code for the barn\n you wish to join"),
+                            TextFormField(
+                              onChanged: (value) {
+                                barnToJoin = value.trim();
+                              },
+                            ),
+                            Row(
+                              children: <Widget>[
+                                Expanded(
+                                  child: ElevatedButton(
+                                    onPressed: () async {
+                                      if (main.MyApp.barnList.isEmpty ||
+                                          main.MyApp.barnList
+                                                  .firstWhere(
+                                                      (element) =>
+                                                          element.id ==
+                                                          barnToJoin,
+                                                      orElse: () => Barn(
+                                                          id: "na",
+                                                          address: "",
+                                                          name: "",
+                                                          ownerid: "",
+                                                          phoneNumber: ""))
+                                                  .id ==
+                                              "na") {
+                                        await linkPersonToBarn(barnToJoin);
+                                      }
+                                      setState(() {});
+                                      Navigator.of(ctx).pop();
+                                    },
+                                    child: const Text(
+                                      'Join Barn',
+                                      style: TextStyle(
+                                          color: Color.fromARGB(
+                                              255, 37, 109, 168)),
+                                    ),
+                                  ),
+                                ),
+                                Expanded(
+                                  child: ElevatedButton(
+                                    onPressed: () {
+                                      Navigator.of(ctx).pop();
+                                    },
+                                    child: const Text(
+                                      'Cancel',
+                                      style: TextStyle(
+                                          color: Color.fromARGB(
+                                              255, 37, 109, 168)),
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            )
+                          ],
+                        ),
+                      );
+                    });
+                setState(() {});
+              }),
+          Divider(
+            height: 1,
+            color: Colors.blueGrey.shade800,
+            thickness: 1,
+          ),
+          ListTile(
+            title: const Text("Logout"),
+            onTap: () async {
+              await FirebaseAuth.instance.signOut();
+
+              main.MyApp.animals.isNotEmpty ? main.MyApp.animals.clear() : {};
+              main.MyApp.barnList.isNotEmpty ? main.MyApp.barnList.clear() : {};
+              main.MyApp.people.isNotEmpty ? main.MyApp.people.clear() : {};
+
+              // return to log in page
+              Navigator.pushReplacement(
+                context,
+                MaterialPageRoute(builder: (context) => LoginPage()),
+              );
             },
           ),
           Divider(
@@ -201,7 +302,8 @@ class _BarnList extends State<BarnList> {
           name: person['name'] ?? '',
           phoneNumber: person['number'] ?? '',
           emergencyPerson: person['emergencyPerson'] ?? '',
-          emergencyNumber: person['emergencyNumber'] ?? '');
+          emergencyNumber: person['emergencyNumber'] ?? '',
+          uid: person['uid'] ?? '');
       return personToReturn;
     } else {
       return Person(
@@ -209,7 +311,8 @@ class _BarnList extends State<BarnList> {
           name: "",
           phoneNumber: "",
           emergencyPerson: "",
-          emergencyNumber: "");
+          emergencyNumber: "",
+          uid: "");
     }
   }
 
@@ -261,5 +364,37 @@ class _BarnList extends State<BarnList> {
     int index =
         main.MyApp.barnList.indexWhere((element) => element.id == barn.id);
     main.MyApp.barnList.removeAt(index);
+  }
+
+  Future<void> linkPersonToBarn(String barnid) async {
+    DocumentReference barnRef =
+        FirebaseFirestore.instance.collection('barns').doc(barnid);
+    DocumentReference personRef = FirebaseFirestore.instance
+        .collection('people')
+        .doc(main.MyApp.currentUser.id);
+
+    DocumentSnapshot barnInfo = await barnRef.get();
+    Map<String, dynamic> barn = barnInfo.data() as Map<String, dynamic>;
+    String ownerid = await getBarnOwner(barn['owner']);
+    main.MyApp.barnList.add(Barn(
+      id: barnRef.id,
+      address: barn['address'] ?? '',
+      name: barn['name'] ?? '',
+      phoneNumber: barn['number'] ?? '',
+      ownerid: ownerid,
+    ));
+
+    await FirebaseFirestore.instance
+        .collection('barn_to_person')
+        .add({'barnid': barnRef, 'personid': personRef});
+  }
+
+  Future<String> getBarnOwner(DocumentReference ownerRef) async {
+    DocumentSnapshot ownerSnapshot = await ownerRef.get();
+    if (ownerSnapshot.exists) {
+      return ownerSnapshot.id;
+    } else {
+      return '';
+    }
   }
 }
